+++ src/components/home/About.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import styled from 'styled-components';
 import PropTypes from 'prop-types';
-import Img from 'gatsby-image';
+
+import PhotoComponent from './PhotoComponent';
 
 const StyledHeading = styled.h1`
   letter-spacing: ;
@@ -28,29 +29,9 @@ const StyledSection = styled.section`
   margin-top: 70px;
 `;
 
-const StyledImageFirst = styled(Img)`
-  max-width: 272px;
-  margin-right: 10px;
-  position: relative !important;
-  transform: translateX(20%);
-`;
-
-const StyledImageSecond = styled(Img)`
-  max-width: 272px;
-  position: relative !important;
-  transform: translateY(-30%);
-`;
-
 const About = ({ data }) => (
   <StyledSection>
-    <StyledImageFirst
-      fluid={data.aboutHouseOne.childImageSharp.fluid}
-      alt="house"
-    />
-    <StyledImageSecond
-      fluid={data.aboutHouseTwo.childImageSharp.fluid}
-      alt="house"
-    />
+    <PhotoComponent data={data} />
     <StyledHeading>about</StyledHeading>
     <StyledDescription>we create delicious memories</StyledDescription>
     <StyledParagraph>
